+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Run 'compact' on each etcd instance before taking the v3 backup.
+- Run 'compact' and 'defrag' on each etcd instance before taking the v3 backup.
 
 ## [2.6.0] - 2021-11-18